+++ crates/db-storage/tests/get_events.rs
@@ -2,7 +2,8 @@ use chrono::{TimeZone as _, Utc};
 use chrono_tz::Tz;
 use database::DbConnection;
 use k3k_db_storage::events::{
-    Event, EventInvite, GetEventsCursor, NewEvent, NewEventInvite, TimeZone,
+    Event, EventId, EventInvite, EventInviteStatus, GetEventsCursor, NewEvent, NewEventInvite,
+    TimeZone, UpdateEventInvite,
 };
 use k3k_db_storage::rooms::{NewRoom, RoomId};
 use k3k_db_storage::users::{NewUser, NewUserWithGroups, UserId};
@@ -31,6 +32,17 @@ fn make_event(conn: &DbConnection, user_id: UserId, room_id: RoomId, hour: u32) 
     }
     .insert(conn)
     .unwrap()
+}
+
+fn update_invite_status(
+    conn: &DbConnection,
+    user_id: UserId,
+    event_id: EventId,
+    new_status: EventInviteStatus,
+) {
+    let changeset = UpdateEventInvite { status: new_status };
+
+    changeset.apply(conn, user_id, event_id).unwrap();
 }
 
 #[tokio::test]
@@ -87,7 +99,8 @@ async fn test() {
 
         // Get first two events 1, 2
         let first_two =
-            Event::get_all_for_user_paginated(&conn, user.id, false, None, None, None, 2).unwrap();
+            Event::get_all_for_user_paginated(&conn, user.id, false, vec![], None, None, None, 2)
+                .unwrap();
         assert_eq!(first_two.len(), 2);
         let query_event1 = &first_two[0].0;
         let query_event2 = &first_two[1].0;
@@ -98,9 +111,17 @@ async fn test() {
         let cursor = GetEventsCursor::from_last_event_in_query(query_event2);
 
         // Use that to get 3,4
-        let next_two =
-            Event::get_all_for_user_paginated(&conn, user.id, false, None, None, Some(cursor), 2)
-                .unwrap();
+        let next_two = Event::get_all_for_user_paginated(
+            &conn,
+            user.id,
+            false,
+            vec![],
+            None,
+            None,
+            Some(cursor),
+            2,
+        )
+        .unwrap();
         assert_eq!(first_two.len(), 2);
         let query_event3 = &next_two[0].0;
         let query_event4 = &next_two[1].0;
@@ -110,9 +131,17 @@ async fn test() {
         // Then 5,6
         let cursor = GetEventsCursor::from_last_event_in_query(query_event4);
 
-        let next_two =
-            Event::get_all_for_user_paginated(&conn, user.id, false, None, None, Some(cursor), 2)
-                .unwrap();
+        let next_two = Event::get_all_for_user_paginated(
+            &conn,
+            user.id,
+            false,
+            vec![],
+            None,
+            None,
+            Some(cursor),
+            2,
+        )
+        .unwrap();
         assert_eq!(first_two.len(), 2);
         let query_event5 = &next_two[0].0;
         let query_event6 = &next_two[1].0;
@@ -122,9 +151,17 @@ async fn test() {
         // Then 7,8
         let cursor = GetEventsCursor::from_last_event_in_query(query_event6);
 
-        let next_two =
-            Event::get_all_for_user_paginated(&conn, user.id, false, None, None, Some(cursor), 2)
-                .unwrap();
+        let next_two = Event::get_all_for_user_paginated(
+            &conn,
+            user.id,
+            false,
+            vec![],
+            None,
+            None,
+            Some(cursor),
+            2,
+        )
+        .unwrap();
         assert_eq!(first_two.len(), 2);
         let query_event7 = &next_two[0].0;
         let query_event8 = &next_two[1].0;
@@ -138,6 +175,7 @@ async fn test() {
             &conn,
             user.id,
             false,
+            vec![],
             Some(Utc.ymd(2020, 1, 1).and_hms(5, 0, 0)),
             None,
             None,
@@ -154,6 +192,7 @@ async fn test() {
             &conn,
             user.id,
             false,
+            vec![],
             None,
             Some(Utc.ymd(2020, 1, 1).and_hms(5, 0, 0)),
             None,
@@ -176,6 +215,7 @@ async fn test() {
             &conn,
             user.id,
             false,
+            vec![],
             Some(Utc.ymd(2020, 1, 1).and_hms(3, 0, 0)),
             Some(Utc.ymd(2020, 1, 1).and_hms(3, 0, 0)),
             None,
@@ -189,6 +229,202 @@ async fn test() {
         assert_eq!(only_event_at_3h[3].0, event6);
         assert_eq!(only_event_at_3h[4].0, event7);
     }
+}
+
+#[tokio::test]
+#[serial]
+async fn get_events_invite_filter() {
+    let db_ctx = test_util::database::DatabaseContext::new(true).await;
+
+    let conn = db_ctx.db.get_conn().unwrap();
+
+    let inviter = make_user(&conn, "Ingo", "Inviter", "inviter");
+    let invitee = make_user(&conn, "Ingrid", "Invitee", "invitee");
+
+    let room = NewRoom {
+        created_by: inviter.id,
+        password: None,
+    }
+    .insert(&conn)
+    .unwrap();
+
+    let accept_event = make_event(&conn, inviter.id, room.id, 1);
+    let decline_event = make_event(&conn, inviter.id, room.id, 1);
+    let tentative_event = make_event(&conn, inviter.id, room.id, 1);
+    let pending_event = make_event(&conn, inviter.id, room.id, 1);
+
+    // Check that the creator of the events gets created events when filtering by `Accepted` invite status
+    let all_events = Event::get_all_for_user_paginated(
+        &conn,
+        inviter.id,
+        false,
+        vec![EventInviteStatus::Accepted],
+        None,
+        None,
+        None,
+        100,
+    )
+    .unwrap();
+
+    assert_eq!(all_events.len(), 4);
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == accept_event.id));
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == decline_event.id));
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == tentative_event.id));
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == pending_event.id));
+
+    // Check that no events are returned when filtering for `Declined`
+    let no_events = Event::get_all_for_user_paginated(
+        &conn,
+        inviter.id,
+        false,
+        vec![EventInviteStatus::Declined],
+        None,
+        None,
+        None,
+        100,
+    )
+    .unwrap();
+
+    assert!(no_events.is_empty());
+
+    let events = vec![
+        &accept_event,
+        &decline_event,
+        &tentative_event,
+        &pending_event,
+    ];
+
+    // invite the invitee to all events
+    for event in events {
+        NewEventInvite {
+            event_id: event.id,
+            invitee: invitee.id,
+            created_by: inviter.id,
+            created_at: None,
+        }
+        .try_insert(&conn)
+        .unwrap();
+    }
+
+    update_invite_status(
+        &conn,
+        invitee.id,
+        accept_event.id,
+        EventInviteStatus::Accepted,
+    );
+
+    update_invite_status(
+        &conn,
+        invitee.id,
+        decline_event.id,
+        EventInviteStatus::Declined,
+    );
+
+    update_invite_status(
+        &conn,
+        invitee.id,
+        tentative_event.id,
+        EventInviteStatus::Tentative,
+    );
+
+    // check `accepted` invites
+    let accepted_events = Event::get_all_for_user_paginated(
+        &conn,
+        invitee.id,
+        false,
+        vec![EventInviteStatus::Accepted],
+        None,
+        None,
+        None,
+        100,
+    )
+    .unwrap();
+
+    assert_eq!(accepted_events.len(), 1);
+    assert!(accepted_events
+        .iter()
+        .any(|(event, ..)| event.id == accept_event.id));
+
+    // check `declined` invites
+    let declined_events = Event::get_all_for_user_paginated(
+        &conn,
+        invitee.id,
+        false,
+        vec![EventInviteStatus::Declined],
+        None,
+        None,
+        None,
+        100,
+    )
+    .unwrap();
+
+    assert_eq!(declined_events.len(), 1);
+    assert!(declined_events
+        .iter()
+        .any(|(event, ..)| event.id == decline_event.id));
+
+    // check `tentative` invites
+    let tentative_events = Event::get_all_for_user_paginated(
+        &conn,
+        invitee.id,
+        false,
+        vec![EventInviteStatus::Tentative],
+        None,
+        None,
+        None,
+        100,
+    )
+    .unwrap();
+
+    assert_eq!(tentative_events.len(), 1);
+    assert!(tentative_events
+        .iter()
+        .any(|(event, ..)| event.id == tentative_event.id));
+
+    // check `pending` invites
+    let pending_events = Event::get_all_for_user_paginated(
+        &conn,
+        invitee.id,
+        false,
+        vec![EventInviteStatus::Pending],
+        None,
+        None,
+        None,
+        100,
+    )
+    .unwrap();
+
+    assert_eq!(pending_events.len(), 1);
+    assert!(pending_events
+        .iter()
+        .any(|(event, ..)| event.id == pending_event.id));
+
+    // expect all events when no invite_status_filter is set
+    let all_events =
+        Event::get_all_for_user_paginated(&conn, invitee.id, false, vec![], None, None, None, 100)
+            .unwrap();
+
+    assert_eq!(all_events.len(), 4);
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == accept_event.id));
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == decline_event.id));
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == tentative_event.id));
+    assert!(all_events
+        .iter()
+        .any(|(event, ..)| event.id == pending_event.id));
 }
 
 #[tokio::test]